rename: 🚚 (mobile) rename theme files
