+++ apps/mobile/src/features/theme/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./useTheme";
+export { useTheme } from "./theme";
